Replace old route pattern with new
